Fix warning: implicit conversion changes signedness
visp/tutorial/particle-filter-curve-fitting/tutorial-pf-curve-fitting-all.cpp:337:136: warning: implicit conversion changes signedness: 'unsigned int' to 'const int' [-Wsign-conversion]
  337 |   lmsFitter.display(data.m_I_orig, vpColor::red, static_cast<unsigned int>(data.m_ipLegend.get_v() + 2 * data.m_legendOffset.get_v()), static_cast<unsigned int>(data.m_ipLegend.get_u()));
      |             ~~~~~~~                                                                                                                    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/tutorial/particle-filter-curve-fitting/tutorial-pf-curve-fitting-all.cpp:337:50: warning: implicit conversion changes signedness: 'unsigned int' to 'const int' [-Wsign-conversion]
  337 |   lmsFitter.display(data.m_I_orig, vpColor::red, static_cast<unsigned int>(data.m_ipLegend.get_v() + 2 * data.m_legendOffset.get_v()), static_cast<unsigned int>(data.m_ipLegend.get_u()));
      |             ~~~~~~~                              ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
visp/tutorial/particle-filter-curve-fitting/tutorial-pf-curve-fitting-all.cpp:528:19: warning: implicit conversion changes signedness: 'long' to 'unsigned long' [-Wsign-conversion]
  528 |     seedPF = data.m_pfSeed;
      |            ~ ~~~~~^~~~~~~~
visp/tutorial/particle-filter-curve-fitting/tutorial-pf-curve-fitting-all.cpp:548:77: warning: implicit conversion changes signedness: 'unsigned long' to 'const long' [-Wsign-conversion]
  548 |   vpParticleFilter<std::vector<vpImagePoint>> filter(nbParticles, stdevsPF, seedPF, nbThread);
      |                                               ~~~~~~                        ^~~~~~
visp/tutorial/particle-filter-curve-fitting/tutorial-pf-curve-fitting-all.cpp:624:91: warning: implicit conversion changes signedness: 'const unsigned int' to 'const int' [-Wsign-conversion]
  624 |     lmsFitter.display<unsigned char>(data.m_IskeletonNoisy, vpColor::gray, legendLmsVert, legendLmsHor);
      |               ~~~~~~~                                                                     ^~~~~~~~~~~~
visp/tutorial/particle-filter-curve-fitting/tutorial-pf-curve-fitting-all.cpp:624:76: warning: implicit conversion changes signedness: 'const unsigned int' to 'const int' [-Wsign-conversion]
  624 |     lmsFitter.display<unsigned char>(data.m_IskeletonNoisy, vpColor::gray, legendLmsVert, legendLmsHor);
      |               ~~~~~~~                                                      ^~~~~~~~~~~~~

+++ tutorial/particle-filter-curve-fitting/tutorial-pf-curve-fitting-pf.cpp
@@ -493,10 +493,10 @@ int main(const int argc, const char *argv[])
   if (returnCode != tutorial::vpTutoCommonData::SOFTWARE_CONTINUE) {
     return returnCode;
   }
-  const unsigned int vertOffset = static_cast<unsigned int>(data.m_legendOffset.get_i());
-  const unsigned int horOffset = static_cast<unsigned int>(data.m_ipLegend.get_j());
-  const int legendPFVert = static_cast<int>(data.m_I_orig.getHeight() - 2 * vertOffset);
-  const int legendPFHor = static_cast<int>(horOffset);
+  const int vertOffset = static_cast<int>(data.m_legendOffset.get_i());
+  const int horOffset = static_cast<int>(data.m_ipLegend.get_j());
+  const int legendPFVert = static_cast<int>(data.m_I_orig.getHeight()) - 2 * vertOffset;
+  const int legendPFHor = horOffset;
 
   // Initialize the attributes of the PF
   //! [Initial_estimates]
